bugfix(convention): Fix `kotlin.UShort incompatible with java.lang.Short` ```console java.lang.ClassCastException: kotlin.UShort incompatible with java.lang.Short at org.hibernate.type.descriptor.java.ShortTypeDescriptor.unwrap(ShortTypeDescriptor.java:15) at org.hibernate.type.descriptor.sql.SmallIntTypeDescriptor$1.doBind(SmallIntTypeDescriptor.java:46) at org.hibernate.type.descriptor.sql.BasicBinder.bind(BasicBinder.java:73) at org.hibernate.type.AbstractStandardBasicType.nullSafeSet(AbstractStandardBasicType.java:276) at org.hibernate.type.AbstractStandardBasicType.nullSafeSet(AbstractStandardBasicType.java:271) at com.example.marvel.convention.jpa.type.YearMonthType.nullSafeSet(YearMonthType.kt:38) at org.hibernate.type.CompositeCustomType.nullSafeSet(CompositeCustomType.java:236) at org.hibernate.persister.entity.AbstractEntityPersister.dehydrate(AbstractEntityPersister.java:2949) at org.hibernate.persister.entity.AbstractEntityPersister.insert(AbstractEntityPersister.java:3246) at org.hibernate.persister.entity.AbstractEntityPersister.insert(AbstractEntityPersister.java:3782) at org.hibernate.action.internal.EntityInsertAction.execute(EntityInsertAction.java:107) at org.hibernate.engine.spi.ActionQueue.executeActions(ActionQueue.java:604) ```

+++ shared-convention/src/main/kotlin/com/example/marvel/convention/jpa/type/YearMonthType.kt
@@ -35,8 +35,8 @@ object YearMonthType : CompositeUserType {
         YearMonth.of(SHORT.nullSafeGet(rs, names[0], session).toUShort().toInt(), BYTE.nullSafeGet(rs, names[1], session).toUByte().toInt())
 
     override fun nullSafeSet(st: PreparedStatement, value: Any, index: Int, session: SharedSessionContractImplementor) {
-        SHORT.nullSafeSet(st, (value as YearMonth).year.toUShort(), index, session)
-        BYTE.nullSafeSet(st, value.monthValue.toUByte(), index + 1, session)
+        SHORT.nullSafeSet(st, (value as YearMonth).year.toUShort().toShort(), index, session)
+        BYTE.nullSafeSet(st, value.monthValue.toUByte().toByte(), index + 1, session)
     }
 
     override fun deepCopy(value: Any): Any = YearMonth.of((value as YearMonth).year, value.monthValue)
